Added timing on the attack
Because the other versions of the code have timers on them, I thought it
would be nice to have it on the C version as well. I think I'm timing
too much, as I'm including the classification of the words, but it's a
decent metric for now. I can get more specific later.